// types.d.ts
export type HasAbortSignal = {
    /**
     * Signal to abort the check.
     */
    signal?: AbortSignal;
};

/**
 * Diagnostics is an analyzing result.
 * It takes one of the following three statuses:
 *
 * - `safe`: The given RegExp is safe.
 * - `vulnerable`: The given RegExp is potentially vulnerable.
 * - `unknown`: An error is occurred on analyzing.
 *   As a result, it is unknown whether the RegExp is safe or vulnerable.
 */
export type Diagnostics = SafeDiagnostics | VulnerableDiagnostics | UnknownDiagnostics;

/**
 * SafeDiagnostics is a diagnostics against a safe RegExp.
 */
export type SafeDiagnostics = {
    source: string;
    flags: string;
    status: 'safe';
    checker: 'automaton' | 'fuzz';
    complexity: SafeComplexity;
};

/**
 * VulnerableDiagnostics is a diagnostics against a vulnerable RegExp.
 */
export type VulnerableDiagnostics = {
    source: string;
    flags: string;
    status: 'vulnerable';
    checker: 'automaton' | 'fuzz';
    attack: AttackPattern;
    complexity: VulnerableComplexity;
    hotspot: Hotspot[];
};

/**
 * UnknownDiagnostics is a diagnostics when an error is occurred on analyzing.
 */
export type UnknownDiagnostics = {
    source: string;
    flags: string;
    status: 'unknown';
    checker?: 'automaton' | 'fuzz';
    error: Error;
};

/**
 * SafeComplexity is a safe complexity.
 */
export type SafeComplexity = {
    type: 'constant' | 'linear' | 'safe';
    summary: string;
    isFuzz: boolean;
};

/**
 * VulnerableComplexity is a vulnerable complexity.
 */
export type VulnerableComplexity = PolynomialComplexity | ExponentialComplexity;
// ...
export type Hotspot = {
    start: number;
    end: number;
    temperature: 'heat' | 'normal';
};

/**
 * Error is a possible error on analyzing.
 */
export type Error = Timeout | Cancel | Unsupported | Invalid | Unexpected;
// ...
export type AttackPattern = {
    pumps: {
        prefix: string;
        pump: string;
        bias: number;
    }[];
    suffix: string;
    base: number;
    string: string;
    pattern: string;
};